refactor: 팔로우 리포지토리 추상화
