+++ src/modules/extensions/user/follow/api/follow.controller.ts
@@ -10,11 +10,11 @@ import {
   UseGuards,
 } from '@nestjs/common';
 import { ApiOperation, ApiQuery, ApiResponse, ApiTags } from '@nestjs/swagger';
-import { FollowService } from './follow.service';
-import { UsersSummaryResDto } from '../../../base/user/dto/response/usersSummaryRes.dto';
+import { FollowService } from '../core/follow.service';
+import { UsersSummaryResDto } from '../../../../base/user/dto/response/usersSummaryRes.dto';
 import { Request as ExpressRequest } from 'express';
-import { PagingParseIntPipe } from '../../../../common/pipes/pagingParseInt.pipe';
-import { JwtAuthGuard } from '../../../../common/guards/jwtAuth.guard';
+import { PagingParseIntPipe } from '../../../../../common/pipes/pagingParseInt.pipe';
+import { JwtAuthGuard } from '../../../../../common/guards/jwtAuth.guard';
 
 @ApiTags('Follow')
 @Controller('follows')
