+++ src/modules/extensions/user/follow/follow.module.ts
@@ -2,10 +2,10 @@ import { TypeOrmModule } from '@nestjs/typeorm';
 import { forwardRef, Module } from '@nestjs/common';
 import { User } from '../../../../entities/user.entity';
 import { Follow } from '../../../../entities/follow.entity';
-import { FollowService } from './follow.service';
-import { FollowRepository } from './follow.repository';
+import { FollowService } from './core/follow.service';
+import { FollowRepository } from './infrastructure/follow.repository';
 import { ToolModule } from '../../../utils/tool/tool.module';
-import { FollowController } from './follow.controller';
+import { FollowController } from './api/follow.controller';
 import { UserModule } from '../../../base/user/user.module';
 import { JwtModule } from '@nestjs/jwt';
 import { AuthModule } from '../../../base/auth/auth.module';
@@ -19,7 +19,7 @@ import { AuthModule } from '../../../base/auth/auth.module';
     forwardRef(() => AuthModule),
   ],
   controllers: [FollowController],
-  providers: [FollowService, FollowRepository],
-  exports: [FollowService, FollowRepository],
+  providers: [FollowService, { provide: 'IFollowRepository', useClass: FollowRepository }],
+  exports: [FollowService, { provide: 'IFollowRepository', useClass: FollowRepository }],
 })
 export class FollowModule {}
